Support image filtering in RPM and generate spec file list.
To support filtering (additional / removed files) the spec file list
needs to be generated from the actual files in the build root. The previous
behaviour also failed when the svg icon wasn't specified.

This commit also removes the application directory from the hook to
calculate the merge root. Packagers such as RPM might move the application
to a different location during customization and this is not tracked.